Add end-to-end tests for Attribute-Value Masking functionality
- Implement save validation tests for self-inclusion failure, direct API rejection of non-held values, and read-only CEL editor state when masked values are present.
- Create simple editor tests to validate masked-chip UI, dirty-form save behavior, and row-remove lockdown for masked rows.
- Introduce shared helper functions for managing masking-related operations, including enabling/disabling feature flags, creating and deleting custom profile attributes, and handling policy creation and management.
